refactored to split the message and voice components of the discord bot up, which is enabling creating a more advanced should_respond check.  also moved the react agent to its own module for easier reuse.  after the refactor, will work on creating tools that are context aware so that they can do things specific to the user or channel that made the request.
